UiHost - to DefaultDriverInstantiation, added setting of implicite timeout to configured 'timeout' value

+++ src/main/java/com/craigstockton/uinavigator/Configuration.java
@@ -6,7 +6,7 @@ import java.util.Properties;
 
 public class Configuration {
 
-    final int timeout;
+    final long timeout;
     private static Configuration config;
 
     private Configuration() {
